+++ src/linear_equations_iterative.cpp
@@ -1,8 +1,9 @@
 #include "linear_equations_iterative.h"
-iterative_method::iterative_method(const vector<vector<double>>& A, const vector<double> &b,double diff):m_A(A),m_B(b),n(b.size()),differ(diff) {}
+iterative_method::iterative_method(const vector<vector<double>>& A, const vector<double> &b, double diff)
+    : m_A(A), m_B(b), n(static_cast<int>(b.size())), differ(diff) {}
 
 void iterative_method::print() const {
-	for (size_t i = 0; i < n; ++i) {
+	for (int i = 0; i < n; ++i) {
 		cout << "x" << i << ":  " << ans[i] << endl;
 	}
 }
